# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SafeBid API
  description: Event-Driven C2C Auction Platform API Contract. Designed around Vibe Code V5 Screen List.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
paths:
...
  /wallet/balance:
    get:
      summary: Get Wallet Balance
      security:
        - cookieAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  availableBalance: { type: number }
                  holdAmount: { type: number }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /wallet/transactions:
    get:
      summary: Get Wallet Transactions
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: OK
        '401': { $ref: '#/components/responses/Unauthorized' }
  /wallet/withdraw:
    post:
      summary: Request Withdrawal
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number }
      responses:
        '200': { description: OK }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: Conflict }

  # Webhooks
...